move Position to ModMapLayer
