+++ patches/tModLoader/Terraria/Map/IMapLayer.TML.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-
-namespace Terraria.Map;
+﻿namespace Terraria.Map;
 
 public partial interface IMapLayer
 {
@@ -11,34 +9,4 @@ public partial interface IMapLayer
 	bool Visible { get; internal set; }
 
 	void Hide() => Visible = false;
-
-	Position GetDefaultPosition() => new Before(null);
-
-	IEnumerable<Position> GetModdedConstraints() => null;
-
-	#region Sort Positions
-
-	public abstract class Position;
-
-	public sealed class Before : Position
-	{
-		public IMapLayer Layer { get; }
-
-		public Before(IMapLayer layer)
-		{
-			Layer = layer;
-		}
-	}
-
-	public sealed class After : Position
-	{
-		public IMapLayer Layer { get; }
-
-		public After(IMapLayer layer)
-		{
-			Layer = layer;
-		}
-	}
-
-	#endregion
 }
